src/webservice/requirements.txt
prefect
tqdm
requests
joblib
pyarrow
sqlalchemy
psycopg2-binary 
evidently 
pandas
numpy
scikit-learn
jupyter
matplotlib
python-dotenv
mlflow
google-cloud-storage
pandas
click
prometheus-fastapi-instrumentator
pydantic
fastapi
mlflow
uvicorn
google-cloud-storage
prometheus-fastapi-instrumentator
python-dotenv
python-multipart
